refactor: schemas 내부 status 폴더 생성 및 이동 (#21)
